flash-tl-image: Add '--profile', '--action' args
This patch allows to choose a profile from the package description
'package.toml' to be applied. The default profile is the first one
found in 'package.toml'.

Each profile contains a set of actions. The default action is "all".
In this case, all actions of the profile will be executed.

The JSON representation of 'package.toml' looks like this:

  {
      "info": {
          "format_version": "0.0.1",
          "description": "List of actions to be done with SPI NOR",
          "build_id": ""
      },
      "profile": {
          "profile-1": {
              "action-1": {
                  "description": <STR>,
                  "command": "flash",
                  "name": <STR>
              },
              "action-2": {
                  "description": <STR>,
                  "command": "erase",
                  "offset": <INT>,
                  "size": <INT>,
                  "negative_offset": true
              },
          },
          "profile-2": {
              "action-1": {
                  "description": <STR>,
                  "command": "flash",
                  "name": <STR>,
                  "offset": {
                      "A": <INT>,
                      "B": <INT>
                  }
              }
          }
      }
  }

The entire list of profiles and actions, can be seen in concrete
`package.toml`.

The 'profile' is selected by passing the argument `--profile`:

  mcom03-flash --port /dev/ttyUSBx flash-tl-image --profile <profile> \
               qspi0 <path_to>/*.tl-image

The 'action' is selected by passing the argument `--action`:

  mcom03-flash --port /dev/ttyUSBx flash-tl-image --profile <profile> \
               --action <action> qspi0 <path_to>/*.tl-image

Issue: #MCOM03SW-2271
Change-Id: Ie19209b791e8ea59bb0fb78cb48aa2083f594e45